Articles API implemented

+++ src/Application/Command/Article/DeleteArticle/DeleteArticleCommandHandler.php
@@ -6,8 +6,10 @@ namespace App\Application\Command\Article\DeleteArticle;
 
 use App\Application\Bus\Command\CommandHandlerInterface;
 use App\Application\Bus\Command\CommandInterface;
+use App\Application\Exception\Article\ArticleAccessDeniedException;
 use App\Domain\Article\Exception\ArticleNotFoundException;
 use App\Domain\Article\Repository\ArticleRepositoryInterface;
+use App\Domain\User\ValueObject\UserRole;
 
 readonly class DeleteArticleCommandHandler implements CommandHandlerInterface
 {
@@ -16,6 +18,9 @@ readonly class DeleteArticleCommandHandler implements CommandHandlerInterface
     ) {
     }
 
+    /**
+     * @throws ArticleAccessDeniedException
+     */
     public function handle(CommandInterface $command): void
     {
         \assert($command instanceof DeleteArticleCommand);
@@ -23,6 +28,16 @@ readonly class DeleteArticleCommandHandler implements CommandHandlerInterface
         $article = $this->articleRepository->findById($command->id);
         if ($article === null) {
             throw ArticleNotFoundException::withId($command->id);
+        }
+
+        if (
+            $command->author->getRole() === UserRole::READER
+            || (
+                $command->author->getRole() === UserRole::AUTHOR
+                && $article->getAuthor()->getId() !== $command->author->getId()
+            )
+        ) {
+            throw new ArticleAccessDeniedException();
         }
 
         $this->articleRepository->remove($article);
